requirement.txt
black==22.12.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
six==1.16.0
Werkzeug==2.2.2
